Add gevent entry point for bridge
