--- judge/bridge/__init__.py
@@ -1,4 +1,0 @@
-from .djangohandler import DjangoHandler
-from .judgecallback import DjangoJudgeHandler
-from .judgehandler import JudgeHandler
-from .judgelist import JudgeList
